search commands modified (BUG: hayoo search for "type:package" gives a single package as result)
in Search command 2 fields added
for filtering the descr fields and the weight
within the result list.
This can be useful e.g. to query the hayoo server for
the package ranks when functions are indexed and the
functions get the same weight as the package containing the function

A Select command added for raw queries without ranking and ordering of the
results.

BUG (not yet located): the index seems to be buggy.
When filling the server with all hackage packages and then
search for type:package just a single package (in alphabetic
ordering the 1.: 4blocks) is returned instead of all packages.