add two new index implementations to list of default implementations
1) SimplePrefixTreeIndex
Like the InvertedIndex, but stores only DocId's as a DocIdSet as Index Values.
There are cases where text search is wanted, but position information is irrelevant.
This implementation can be used for this cases to save some memory space

2) Geospacial index
Both the implementation based on the InvertedIndex and based on the RTree are now
in the list of default implementations.