in hunt-searchengine IntermediateValue refactored into a sum type SearchResult
class IndexValue simplified, writing into indexes is done via Occurrences,
this is the most general data type for describing word occurrences with
a set of positions. Reading (computing search result) is done via Searchresult type.
This is a sum type with 3 variants, the most general Occurrences variant,
the least general, just a set of DocId's and in between a set of DocId's with an
associated Score e.g acting as a frequency count.

The old RawResult type (and module) then becomes redundant and is removed.
The Intermediate module only contains the class IndexValue and the instances
for this class. That module is tightly coupled with module Index, so it's
moved into module Index.

The scoring has become a bit more complex, so there are two extra
modules for scoring, Hunt.Scoring.Score defines the Score type the class
ScoredResult an basic instances of this class.
Hunt.Scoring.SearchResult contains the SearchResult type and its
instance of ScoredResult. The real perfomance gain with SearchResult
is implemented in this instance. Combining search results of the
three variants is optimized by a case analysis to eliminate unneccessary
conversions. Also conversions of SearchResult into more aggregated and simpler
types is optimized.