Split ContextIndex into segments
While the ongoing approach so far maintained an 'active' ContextMap + DocTable
and separated snapshots, this try takes us to new places in our
experiments.

The basic idea: The ContextIndex consists of segments. There is no
separation. Any words+docs added to the Contextindex are put in its own
Segment. Once a Segment is created its ContextMap and DocTable can not
be modified. Removals of DocIds and Contexts are represented in separate
Sets. Since this has the potential to float the memory the Segments have
to be merged together perdiodically to a) garbage collect deleted
documents and contexts  and b) reduce the overall amount of segments
since lookup time is now dependend on the number of segments in a
ContextIndex.